Updated dependencies

+++ include/nrpe/packet.hpp
@@ -108,7 +108,7 @@ namespace nrpe {
 		packet(const char *buffer, unsigned int buffer_length) : tmpBuffer(NULL), payload_length_(length::get_payload_length(buffer_length)) {
 			readFrom(buffer, buffer_length);
 		};
-		packet(short type, short version, int result, std::string payLoad, unsigned int payload_length)
+		packet(short type, short version, int16_t result, std::string payLoad, unsigned int payload_length)
 			: tmpBuffer(NULL)
 			, payload_length_(payload_length)
 			, type_(type)
